Merge upload_max_newlines with upstream.
Slightly different implementation than Suhosin 0.9.38-dev as PHP 5.4.41 fix had been merged.

Results with this new set of fixes are quite good: Takes about 0.005 seconds to reject a malious 2M upload, in comparison with PHP 5.4.41 it took 0.1 seconds.

Without any of the fixes this would lead to a DoS condition.