Improve performance of post_handler in PHP 5.6
BUFSIZ is a macro that is defined by the compiler and in Visual Studio its value is 512 while in other systems values range from 1024 to 4096. This causes a performance penalty while parsing POST that sometimes is rather serious, up to 32 times slower!

To fix this, BUFSIZ has been replaced with SAPI_POST_BLOCK_SIZE, a #define found in SAPI.h that is 0x4000 (16384) in PHP 5.6. This change will make POST buffer use exactly the same size for both read and write operations, thus both operations will have similar timings.